Project 6. SqLite III 1- Generate onListItemClick and call the form activity through an intent, and indicate how display and record identifier
2- In the main we have previously declared two constants to edit mode and display mode.

Introduce the id in the parameters that matches the id of Bd

start the activity with startActivityForResult (i, C_VISUALIZAR);

3-Create the form:
In the layout insert a scrollview and within one linnearLayout two Textviews with label id and label name
and EditText for the name you edit or not.

4-In the Activity form will capture data using a tried in a bundle class, not extracted

directly through the intent, if we store for later extraction.

5-generate the query again because we will consult a data